auth with personal access tokens

+++ app/Http/Controllers/V1/AlbumController.php
@@ -7,15 +7,16 @@ use App\Http\Requests\StoreAlbumRequest;
 use App\Http\Requests\UpdateAlbumRequest;
 use App\Http\Resources\V1\AlbumResource;
 use App\Models\Album;
+use Illuminate\Http\Request;
 
 class AlbumController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        return AlbumResource::collection(Album::paginate());
+        return AlbumResource::collection(Album::where('user_id',$request->user()->id)->paginate());
     }
 
     /**
@@ -23,15 +24,21 @@ class AlbumController extends Controller
      */
     public function store(StoreAlbumRequest $request)
     {
-        $album =  Album::create($request->all());
+
+        $data = $request->all();
+        $data['user_id'] = $request->user()->id;
+        $album =  Album::create($data);
         return new AlbumResource($album);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Album $album)
+    public function show(Request $request,Album $album)
     {
+        if($request->user()->id != $album->user_id){
+            return abort(403,'Unauthorized');
+        }
         return new AlbumResource($album);
     }
 
@@ -40,6 +47,9 @@ class AlbumController extends Controller
      */
     public function update(UpdateAlbumRequest $request, Album $album)
     {
+        if($request->user()->id != $album->user_id){
+            return abort(403,'Unauthorized');
+        }
         $album->update($request->all());
 
         return new AlbumResource($album);
@@ -48,8 +58,11 @@ class AlbumController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Album $album)
+    public function destroy(Request $request,Album $album)
     {
+        if($request->user()->id != $album->user_id){
+            return abort(403,'Unauthorized');
+        }
         $album->delete();
 
         return response('',204);
