+++ app/Http/Controllers/V1/ImageManipulationController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateImageManipulationRequest;
 use App\Http\Resources\V1\ImageManipulationResource;
 use App\Models\Album;
 use App\Models\ImageManipulation;
+use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
@@ -20,13 +21,16 @@ class ImageManipulationController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        return ImageManipulationResource::collection(ImageManipulation::paginate());
+        return ImageManipulationResource::collection(ImageManipulation::where('user_id',$request->user()->id)->paginate());
     }
 
-    public function byAlbum(Album $album)
+    public function byAlbum(Request $request,Album $album)
     {
+        if($request->user()->id != $album->user_id){
+            return abort(403,'Unauthorized');
+        }
         $where = [
             'album_id' => $album->id
         ];
@@ -47,10 +51,16 @@ class ImageManipulationController extends Controller
         $data = [
             'type' => ImageManipulation::TYPE_RESIZE,
             'data' => json_encode($all),
-            'user_id' => null,
+            'user_id' => $request->user()->id,
         ];
 
         if (isset($all['album_id'])) {
+            $album = Album::find($all['album_id']);
+
+            if($request->user()->id != $album->user_id){
+                return abort(403,'Unauthorized');
+            }
+
             $data['album_id'] = $all['album_id'];
         }
 
@@ -98,8 +108,11 @@ class ImageManipulationController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(ImageManipulation $image)
+    public function show(Request $request,ImageManipulation $image)
     {
+        if($request->user()->id != $image->user_id){
+            return abort(403,'Unauthorized');
+        }
         return new ImageManipulationResource($image);
     }
 
@@ -107,8 +120,11 @@ class ImageManipulationController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(ImageManipulation $image)
+    public function destroy(Request $request,ImageManipulation $image)
     {
+        if($request->user()->id != $image->user_id){
+            return abort(403,'Unauthorized');
+        }
         $image->delete();
         return response('',204);
     }
